requirements.txt
# Multimodal Rare Disease Diagnosis Framework
# Core Dependencies

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers (BioBERT, ClinicalBERT)
transformers>=4.30.0
tokenizers>=0.13.0

# Computer Vision
opencv-python>=4.7.0
Pillow>=9.5.0
facenet-pytorch>=2.5.0  # MTCNN face detection

# Advanced Image Augmentation (for limited data scenarios)
albumentations>=1.3.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dateutil>=2.8.0

# XML Parsing (for Orphadata)
lxml>=4.9.0

# Optional: Jupyter for explainability notebook
jupyter>=1.0.0
ipywidgets>=8.0.0

# Optional: Streamlit for demo UI
# streamlit>=1.22.0

# Development
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0
